migrated mui5 on helm ui and moved common class between helm and devfile ui
