+++ src/webview/common/loading.style.tsx
@@ -3,17 +3,18 @@
  *  Licensed under the MIT License. See LICENSE file in the project root for license information.
  *-----------------------------------------------------------------------------------------------*/
 
-import { Theme, createStyles } from '@material-ui/core/styles';
+import { Theme } from '@mui/material/styles';
+import { createStyles } from '@mui/styles';
 
 export default (_theme: Theme) =>
     createStyles({
-        devfileGalleryGrid: {
-            margin: '1.5rem 0rem',
-            padding: '0rem 1.5rem',
-            color: 'var(--vscode-foreground)',
-            backgroundColor: 'var(--vscode-editor-background)',
-            fontSize: 'var(--vscode-font-size)',
-            fontFamily: 'var(--vscode-font-family)',
-            zIndex: 0
+        loadProgress: {
+            color: '#EE0000'
         },
+        loading: {
+            display: 'flex',
+            justifyContent: 'center',
+            alignItems: 'center',
+            minHeight: '100vh'
+        }
     });
