+++ src/webview/common/searchBar.style.tsx
