+++ src/webview/common/starterProjectDisplay.style.tsx
