--- src/webview/devfile-registry/app/cardItemStyle.scss
@@ -1,7 +1,0 @@
-/*-----------------------------------------------------------------------------------------------
- *  Copyright (c) Red Hat, Inc. All rights reserved.
- *  Licensed under the MIT License. See LICENSE file in the project root for license information.
- *-----------------------------------------------------------------------------------------------*/
-input[type='text'] {
-    padding: 10px 0 !important
-}
