+++ src/webview/devfile-registry/app/filterElements.tsx
@@ -3,21 +3,11 @@
  *  Licensed under the MIT License. See LICENSE file in the project root for license information.
  *-----------------------------------------------------------------------------------------------*/
 
-import React, { ChangeEvent } from 'react';
-import { DefaultProps } from './home';
-import {
-    Registry
-} from './../../../odo/componentType';
+import { Checkbox, FormControlLabel, Tooltip, Typography } from '@mui/material';
 import { makeStyles } from '@mui/styles';
+import React from 'react';
+import { FilterProps } from '../../common/propertyTypes';
 import filterElementsStyle from './filterElements.style';
-import { Checkbox, FormControlLabel } from '@material-ui/core';
-import { Tooltip, Typography } from '@mui/material';
-
-interface FilterProps extends DefaultProps {
-    id: string,
-    registries: Registry[],
-    onCheckBoxChange: (event: ChangeEvent<HTMLInputElement>, checked: boolean) => void;
-}
 
 const filterStyle = makeStyles(filterElementsStyle);
