--- src/webview/helm-chart/app/errorPage.style.tsx
@@ -1,20 +1,0 @@
-/*-----------------------------------------------------------------------------------------------
- *  Copyright (c) Red Hat, Inc. All rights reserved.
- *  Licensed under the MIT License. See LICENSE file in the project root for license information.
- *-----------------------------------------------------------------------------------------------*/
-
-import { Theme, createStyles } from '@material-ui/core/styles';
-
-export default (_theme: Theme) =>
-    createStyles({
-        error: {
-            position: 'absolute',
-            top: '50%',
-            left: '50%',
-            transform: 'translate(-50%, -50%)',
-            fontSize: 'large',
-            color: 'Red',
-            border: '2px solid var(--vscode-focusBorder)',
-            padding: '1.5rem'
-        }
-    });
