+++ src/webview/helm-chart/app/wrapperCardItem.tsx
@@ -11,7 +11,6 @@ export interface DevFileProps extends React.AnchorHTMLAttributes<HTMLAnchorEleme
     helmEntry: ChartResponse;
     cardItemStyle: any;
     themeKind: number;
-    projectDisplayStyle: any;
 }
 
 export const WrapperCardItem: React.ForwardRefExoticComponent<DevFileProps> = forwardRef((props, ref) => <CardItem {...props} />)
