+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
     "compilerOptions": {
+        "jsx": "react",
         "module": "commonjs",
         "target": "es6",
         "outDir": "out",
